size_trunc_len(): drastically simplify function by applying autoscale from size_trunc()
This function was used to scale according to the supplied length. Now, it will
scale at most to 8 characters by passing the size to size_trunc() and appending
to the supplied buffer.

It is important to note that really small buffers have to be bigger than it
appears because they store bytes and non-Latin scripts need more than one byte
with UTF-8 to encode them, e.g., the string "1023 МиБ" in Russian requires
11 bytes + null terminator.